[Graphics] Graphics buffers can now allocate sub-buffers
[Graphics] Implemented GraphicsBuffer.Allocate()
[Graphics] Added GraphicsBuffer.Offset property
[Graphics] Added GraphicsBuffer.AllocatedBytes property
[Graphics] Added GraphicsBuffer.Alignment property
[Graphics] Added GraphicsBuffer.ParentBuffer property
[Graphics] GraphicsBuffer.GetData() and SetData() now respect buffer offset

+++ Molten.Engine/Graphics/Pipeline/GraphicsQueue.cs
@@ -160,12 +160,12 @@ public abstract class GraphicsQueue : EngineObject
     /// <param name="mapType">The type of mapping to perform.</param>
     /// <returns></returns>
     /// <exception cref="GraphicsResourceException"></exception>
-    public unsafe GraphicsStream MapResource(GraphicsResource resource, uint subresource, uint offsetBytes, GraphicsMapType mapType)
+    public unsafe GraphicsStream MapResource(GraphicsResource resource, uint subresource, ulong offsetBytes, GraphicsMapType mapType)
     {
         resource.Ensure(this);
         ResourceMap map = GetResourcePtr(resource, subresource, mapType);
         GraphicsStream stream = new GraphicsStream(this, resource, ref map);
-        stream.Position = offsetBytes;
+        stream.Position = (long)offsetBytes;
         return stream;
     }
